Added efficient address_recorder
Why: This class allows the program to very quickly insert
entires for hit addresses during replay. This should minimize
performance impact. It also allows for quick searching of the data
between any two timestamps, and allows for the user to
re-run a frametime and refresh all of the values inside
without having to touch anything else

How: This uses a tired approach of a vec of bins. Each framtime
gets a bin that holds all of the timestamps that happned
inside of it and all of the addresses that were reached during
the execution of the frametime.

Notes: This object is not threadsafe but I haven't added the
!Send trait because I hate the joke that is PhantomData.