MVP that supports creating dot data from trace
Why: Because I need something for tomorrow and I don't have time to do
it properly.

How: I cheated a bit with the graph builder and I have a bug with
librr_rs resumes. For some reason, resuming from an event causes
an infinite loop. I don't know why I dont have a test for that